Vertx web - make hot replacement work
