+++ core/devmode/src/main/java/io/quarkus/dev/RuntimeUpdatesProcessor.java
@@ -78,14 +78,15 @@ public class RuntimeUpdatesProcessor implements HotReplacementContext {
         return DevModeMain.deploymentProblem;
     }
 
-    public void doScan() throws IOException {
+    public boolean doScan() throws IOException {
         final long startNanoseconds = System.nanoTime();
         final ConcurrentMap<String, byte[]> changedClasses = scanForChangedClasses();
         if (changedClasses == null)
-            return;
+            return false;
 
         DevModeMain.restartApp();
         log.infof("Hot replace total time: %ss ", Timing.convertToBigDecimalSeconds(System.nanoTime() - startNanoseconds));
+        return true;
     }
 
     ConcurrentMap<String, byte[]> scanForChangedClasses() throws IOException {
