+++ extensions/vertx/deployment/src/main/java/io/quarkus/vertx/deployment/VertxBuildItem.java
@@ -1,0 +1,20 @@
+package io.quarkus.vertx.deployment;
+
+import org.jboss.builder.item.SimpleBuildItem;
+
+import io.quarkus.runtime.RuntimeValue;
+import io.vertx.core.Vertx;
+
+public final class VertxBuildItem extends SimpleBuildItem {
+
+    private final RuntimeValue<Vertx> vertx;
+
+    public VertxBuildItem(RuntimeValue<Vertx> vertx) {
+        this.vertx = vertx;
+    }
+
+    public RuntimeValue<Vertx> getVertx() {
+        return vertx;
+    }
+
+}
